@@ -6,13 +6,13 @@ import 'src/mocks.dart';
 
 void main() {
   group('MergeRequestsApi', () {
-    MockGitLabHttpClient mockHttpClient;
+    late MockGitLabHttpClient mockHttpClient;
     GitLab gitLab;
-    ProjectsApi project;
+    late ProjectsApi project;
 
     final projectId = 1337;
 
-    final mergeRequestsMap = data.decodeMap(data.mergeRequest);
+    final mergeRequestsMap = data.decodeMap(data.mergeRequest)!;
     final mergeRequestId = mergeRequestsMap['id'] as int;
     final mergeRequestIid = mergeRequestsMap['iid'] as int;
 
@@ -36,7 +36,7 @@ void main() {
       expect(mergeRequest.title, mergeRequestsMap['title']);
       expect(mergeRequest.state, mergeRequestsMap['state']);
 
-      expect(mergeRequest.author.id, mergeRequestsMap['author']['id']);
+      expect(mergeRequest.author!.id, mergeRequestsMap['author']['id']);
       expect(
           mergeRequest.assignees[0].id, mergeRequestsMap['assignees'][0]['id']);
 
