@@ -6,12 +6,12 @@ import 'src/mocks.dart';
 
 void main() {
   group('ReleasesApi', () {
-    MockGitLabHttpClient mockHttpClient;
+    late MockGitLabHttpClient mockHttpClient;
     GitLab gitLab;
-    ProjectsApi project;
+    late ProjectsApi project;
 
     final projectId = 1337;
-    final releasesJson = data.decodeList(data.releasesList);
+    final releasesJson = data.decodeList(data.releasesList)!;
     final releaseTag = releasesJson[1]["tag_name"] as String;
 
     setUp(() {
@@ -20,7 +20,7 @@ void main() {
       project = gitLab.project(projectId);
     });
 
-    List<Map<String, dynamic>> fromJsonList(jsonList) => jsonList is List
+    List<Map<String, dynamic>?>? fromJsonList(jsonList) => jsonList is List
         ? jsonList
             .map((jl) => jl is Map<String, dynamic> ? jl : null)
             .where((jl) => jl != null)
@@ -70,9 +70,9 @@ void main() {
 
         expect(release.milestones?.length, milestonesJson?.length);
         if (release.milestones != null) {
-          for (var i = 0; i < release.milestones.length; i++) {
-            final Milestone milestone = release.milestones[i];
-            final milestoneJson = milestonesJson[i];
+          for (var i = 0; i < release.milestones!.length; i++) {
+            final Milestone milestone = release.milestones![i]!;
+            final milestoneJson = milestonesJson![i]!;
 
             expect(milestone.id, milestoneJson['id']);
             expect(milestone.iid, milestoneJson['iid']);
@@ -82,7 +82,7 @@ void main() {
             expect(
                 milestone.state,
                 enumFromString(
-                    MilestoneState.values, milestoneJson['state'] as String));
+                    MilestoneState.values, milestoneJson['state'] as String?));
             expect(milestone.createdAt,
                 DateTime.parse(milestoneJson['created_at'].toString()));
             expect(milestone.updatedAt,
@@ -96,24 +96,24 @@ void main() {
         }
 
         final Asset asset = release.assets;
-        final sourcesJson = fromJsonList(assetsJson["sources"]);
-        final linksJson = fromJsonList(assetsJson["links"]);
+        final sourcesJson = fromJsonList(assetsJson["sources"])!;
+        final linksJson = fromJsonList(assetsJson["links"])!;
 
         expect(asset.count, assetsJson['count']);
 
-        expect(asset.sources.length, sourcesJson.length ?? 0);
-        for (var i = 0; i < asset.sources.length; i++) {
-          final Source source = asset.sources[i];
-          final sourceJson = sourcesJson[i];
+        expect(asset.sources!.length, sourcesJson.length);
+        for (var i = 0; i < asset.sources!.length; i++) {
+          final Source source = asset.sources![i]!;
+          final sourceJson = sourcesJson[i]!;
 
           expect(source.format, sourceJson['format']);
           expect(source.url, sourceJson['url']);
         }
 
-        expect(asset.links.length, linksJson.length);
-        for (var i = 0; i < asset.links.length; i++) {
-          final Link link = asset.links[i];
-          final linkJson = linksJson[i];
+        expect(asset.links!.length, linksJson.length);
+        for (var i = 0; i < asset.links!.length; i++) {
+          final Link link = asset.links![i]!;
+          final linkJson = linksJson[i]!;
 
           expect(link.id, linkJson['id']);
           expect(link.name, linkJson['name']);
@@ -147,8 +147,8 @@ void main() {
 
       call.verifyCalled(1);
       expect(releases, hasLength(2));
-      expect(releases[0].tagName, "v0.2");
-      expect(releases[1].tagName, "v0.1");
+      expect(releases[0]!.tagName, "v0.2");
+      expect(releases[1]!.tagName, "v0.1");
     });
 
     test('.createFromTag(v0.3)', () async {
