@@ -6,8 +6,8 @@ export 'json_issue_notes_data.dart';
 export 'json_merge_requests_data.dart';
 export 'json_releases_data.dart';
 
-Map decodeMap(String json) => jsonDecode(json) as Map;
-List decodeList(String json) => jsonDecode(json) as List;
+Map? decodeMap(String json) => jsonDecode(json) as Map?;
+List? decodeList(String json) => jsonDecode(json) as List?;
 
 const job = '''
 {
